refactor: user type
