+++ prisma/schema.prisma
@@ -13,10 +13,15 @@ datasource db {
 model User {
   id             String            @id @default(auto()) @map("_id") @db.ObjectId
 
-  userName String? @unique
-  email    String? @unique
+  particleUUID String? @unique
 
-  name         String?
+  userName String? @unique
+  email    String @unique
+
+  name         String
+  // firstName String
+  // lastName String
+
   profileImage String?
 
   createdAt    DateTime          @default(now()) @db.Timestamp
