+++ src/lib/global_types.ts
@@ -1,6 +1,1 @@
-export type userType = {
-  name: string
-  email: string
-  profileImage: string
-  userName: string
-}
+
